refactor(notifications): drop owner_ts u64 cast; atomic push-token cleanups
Review feedback on #725:
- owner_ts canonical encoding now uses i64::to_be_bytes directly instead of an
  `as u64` cast. The bytes are identical for a non-negative ms timestamp, so
  signatures are unchanged; owner_ts stays i64 to match the codebase timestamp
  convention and SQLite INTEGER.
- remove_group_push_tokens_for_member and remove_stale_group_push_tokens each
  issue two related DELETEs (live rows + tombstones); wrap both in
  with_transaction so the pair is atomic, matching delete_local_group_data and
  the apply_* methods.

+++ crates/marmot-app/src/notifications.rs
@@ -512,7 +512,10 @@ fn push_signed_record_bytes(
     buf.push(platform_byte);
     buf.extend_from_slice(&server_pubkey);
     buf.extend_from_slice(&fingerprint);
-    buf.extend_from_slice(&(owner_ts as u64).to_be_bytes());
+    // owner_ts is i64 throughout (codebase timestamp convention + SQLite INTEGER).
+    // For a non-negative millisecond timestamp its two's-complement big-endian
+    // bytes are identical to the spec's `u64` encoding, so no cast is needed.
+    buf.extend_from_slice(&owner_ts.to_be_bytes());
     buf.extend_from_slice(&relay_len.to_be_bytes());
     buf.extend_from_slice(relay_bytes);
     if let Some(token) = encrypted_token {
